roachtest: disable/patch roachtests for IBM Cloud
This patch updates some tests to make them compatible with IBM Cloud and
the s390x architecture, and disables some others.

Most of the tests that were disabled on Azure were disabled on IBM.

Mixed-version and version upgrade tests have also been disabled, as
s390x is currently only built on CockroachDB's master branch.

Other disabled tests depend on dependencies unavailable on s390x:
- db_console (cypress)
- cdc (kafka)

For tests using a local Prometheus cluster, as Grafana is unavailable on
s390x, the local Grafana is not installed.

Docker images have been built in the testeng Docker repository for
`fluent-bit` and `minio`, and binaries are available for `prometheus`,
`node_exporter` and `ebpf_exporter`.

Jepsen's .jar has been rebuilt to run on Java 17 as the previously
in-use Java 8 version was timeouting during the Analyzis step.

Epic: none

Release note: none

+++ pkg/cmd/roachtest/tests/django.go
@@ -148,6 +148,17 @@ func registerDjango(r registry.Registry) {
 
 		if err := repeatRunE(
 			ctx, t, c, node, "install django's dependencies", `
+				if [ $(arch) == "s390x" ]; then
+					# s390x doesn't have a prebuilt wheel for bcrypt, so we need
+					# to build it from source. This requires rust.
+					# Install rust and set the default toolchain to stable.
+					curl https://sh.rustup.rs -sSf | sh -s -- -y
+					source $HOME/.cargo/env
+
+					# s390x doesn't have a prebuilt wheel for pillow, so we need
+					# to build it from source. This requires libjpeg-dev.
+					sudo apt-get install -y libjpeg-dev
+				fi
 				source venv/bin/activate &&
 				cd /mnt/data1/django/tests &&
 				pip3 install -e .. &&
